requirements.txt
dataclass-wizard
pyyaml
structlog
pytest
